Extract common logging code to pipe commons library
