+++ workflows/pipe-common/scripts/cluster_wait_for_workers.py
@@ -16,43 +16,17 @@ import argparse
 import os
 import subprocess
 import time
-import traceback
 
-from pipeline import Logger, TaskStatus, PipelineAPI, StatusEntry
+from pipeline import Logger, LoggerTask, TaskStatus, PipelineAPI, StatusEntry
 from pipeline import Kubernetes
 
 
-class Task:
-
-    def __init__(self, task_name):
-        self.task_name = task_name
-
-    def fail_task(self, e):
-        Logger.fail('Error in task {}: {}: {}.'
-                    .format(self.task_name, e, traceback.format_exc()), task_name=self.task_name)
-
-    def warn_task(self, e):
-        Logger.warn('Warning in task {}: {}: {}'
-                    .format(self.task_name, e, traceback.format_exc()), task_name=self.task_name)
-
-
-def fail_task_on_error(f):
-    def wrapper(*args, **kwargs):
-        self = args[0]
-        try:
-            return f(*args, **kwargs)
-        except Exception as e:
-            self.fail_task(e)
-            raise
-    return wrapper
-
-
-class CreateWorkerNodes(Task):
+class CreateWorkerNodes(LoggerTask):
 
     INIT_WORKER_TASK = 'InitializeWorker'
 
     def __init__(self):
-        Task.__init__(self, task_name='WaitForWorkerNodes')
+        LoggerTask.__init__(self, task_name='WaitForWorkerNodes')
         self.kube = Kubernetes()
         self.pipe_api = PipelineAPI(os.environ['API'], 'logs')
 
@@ -60,7 +34,7 @@ class CreateWorkerNodes(Task):
         child_runs = self.pipe_api.load_child_pipelines(parent_id)
         return [ c["id"] for c in child_runs ]
 
-    @fail_task_on_error
+    @LoggerTask.fail_on_error
     def await_workers_start(self, nodes_number, parent_id):
         if nodes_number == 0:
             Logger.success('No workers requested. Processing will run on a master node', task_name=self.task_name)
@@ -101,13 +75,13 @@ class CreateWorkerNodes(Task):
         return started
 
 
-class BuildHostfile(Task):
+class BuildHostfile(LoggerTask):
 
     def __init__(self):
-        Task.__init__(self, task_name='BuildHostfile')
+        LoggerTask.__init__(self, task_name='BuildHostfile')
         self.kube = Kubernetes()
 
-    @fail_task_on_error
+    @LoggerTask.fail_on_error
     def run(self, worker_pods, path, run_id):
         Logger.info('Creating hostfile {}'.format(path), task_name=self.task_name)
         with open(path, 'w') as file:
@@ -128,12 +102,12 @@ class BuildHostfile(Task):
         self.execute_command(cmd)
 
 
-class ShutDownCluster(Task):
+class ShutDownCluster(LoggerTask):
 
     def __init__(self):
-        Task.__init__(self, task_name='ShutDownCluster')
+        LoggerTask.__init__(self, task_name='ShutDownCluster')
 
-    @fail_task_on_error
+    @LoggerTask.fail_on_error
     def run(self, worker_ids, status):
         Logger.info('Shutting down {} node(s)'.format(len(worker_ids)), task_name=self.task_name)
         api = PipelineAPI(os.environ['API'], 'logs')
